# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "Advanced Cybersecurity Analytics Service (ACAS) for AI-based analysis with explainable intelligence of network events captured by the Montimage Monitoring Tools (MMT)",
    "termsOfService" : "https://github.com/Montimage/acas",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.eu"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    },
    {
      "description": "Production Server",
      "url": "https://ndr.montimage.eu/api/"
    }
  ],
...
  "paths" : {
...
    "/retrain/" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "The retraining model process status",
        "description" : "Get the status of retraining model process",
        "operationId" : "getRetrainStatus",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/retrainStatus"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "models" ],
        "summary" : "Retraining models using different training datasets",
        "description" : "Retraining models using different training datasets",
        "operationId" : "requestRetrainModel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/retrainConfig"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/retrainStatus"
                }
              }
            }
          },
          "500" : {
            "description" : "Failed to process retraining a model"
          }
        }
      }
    },
    "/retrain/{modelId}" : {
      "post" : {
        "tags" : [ "models" ],
        "summary" : "Retraining a specific model using different training datasets",
        "description" : "Retraining a specific model using different training datasets",
        "operationId" : "requestRetrainADModel",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model file to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/retrainADConfig"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/retrainStatus"
                }
              }
            }
          },
          "500" : {
            "description" : "Failed to process retraining a model"
          }
        }
      }
    },
    "/retrain/offline" : {
      "post" : {
        "tags" : [ "models" ],
        "summary" : "Queue-based offline retraining",
        "description" : "Retrain model using job queue for non-blocking execution",
        "operationId" : "retrainOffline",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "modelId" : {
                    "type" : "string"
                  },
                  "trainingDataset" : {
                    "type" : "string"
                  },
                  "testingDataset" : {
                    "type" : "string"
                  },
                  "training_parameters" : {
                    "$ref" : "#/components/schemas/trainingParameters"
                  },
                  "useQueue" : {
                    "type" : "boolean",
                    "example" : false
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "jobId" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/retrain/job/{jobId}" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get queued retrain job status",
        "description" : "Get the status of a queued retrain job",
        "operationId" : "getRetrainJobStatus",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "description" : "Job ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/jobStatus"
                }
              }
            }
          }
        }
      }
    },
    "/retrain/predictions/{retrainId}" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get predictions from retrain",
        "description" : "Get predictions generated during retraining",
        "operationId" : "getRetrainPredictions",
        "parameters" : [ {
          "name" : "retrainId",
          "in" : "path",
          "description" : "Retrain ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/predictions/{predictionId}" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Get the statistics of prediction",
        "description" : "Get the statistics of prediction where there are total number of flows, number of attack flows, and number of normal flows",
        "operationId" : "getPredictionStats",
        "parameters" : [ {
          "name" : "predictionId",
          "in" : "path",
          "description" : "Name of the model file to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "4bbfaa76-9e39-49b8-baa7-c4a7cdc9fed8"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "prediction" : {
                      "type" : "array",
                      "items" : {
                        "type" : "number",
                        "example" : 1243
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "trainingParameters" : {
        "type" : "object",
        "properties" : {
          "nb_epoch_cnn" : {
            "type" : "number",
            "example" : 5
          },
          "nb_epoch_sae" : {
            "type" : "number",
            "example" : 3
          },
          "batch_size_cnn" : {
            "type" : "number",
            "example" : 16
          },
          "batch_size_sae" : {
            "type" : "number",
            "example" : 32
          }
        }
      },
...
      "retrainStatus" : {
        "type" : "object",
        "properties" : {
          "isRunning" : {
            "type" : "boolean",
            "example" : true
          },
          "config" : {
            "$ref" : "#/components/schemas/retrainConfig"
          },
          "lastRetrainId" : {
            "type" : "string",
            "example" : "422e5a74-6b14-4ced-a630-b777fbfbb50c"
          },
          "lastRetrainAt" : {
            "type" : "string",
            "example" : "1694699830518"
          }
        }
      },
      "retrainADConfig": {
        "type": "object",
        "properties": {
          "retrainADConfig" : {
            "type" : "object",
            "properties" : {
              "trainingDataset": {
                "type": "string",
                "example": "Train_samples.csv"
              },
              "testingDataset": {
                "type": "string",
                "example": "Test_samples.csv"
              },
              "training_parameters": {
                "$ref": "#/components/schemas/trainingParameters"
              }
            }
          }
        }
      },
      "retrainConfig" : {
        "type" : "object",
        "properties" : {
          "retrainConfig" : {
            "type" : "object",
            "properties" : {
              "modelId" : {
                "type" : "string",
                "example" : "e0a0588e-96c4-4fdf-9237-a8e94a6f5ffe.h5"
              },
              "trainingDataset" : {
                "type" : "string",
                "example" : "/tmp/Botnet_Train.csv"
              },
              "testingDataset" : {
                "type" : "string",
                "example" : "/tmp/Botnet_Test.csv"
              },
              "training_parameters" : {
                "$ref" : "#/components/schemas/trainingParameters"
              }
            }
          }
        }
      },
...
      "jobStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["completed", "failed", "processing", "waiting"],
            "example": "completed"
          },
          "result": {
            "type": "object",
            "description": "Job result data when completed"
          },
          "failedReason": {
            "type": "string",
            "description": "Error message if job failed"
          },
          "progress": {
            "type": "number",
            "description": "Job progress percentage (0-100)"
          }
        }
      }
    }
  }
}